requirements.txt
python-dotenv
groq
PyYAML
weasyprint
Jinja2